pyproject.toml
[project]
name = "chat-migracion"
version = "0.1.0"
description = "Chat bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.30",
    "langchain>=0.3.14",
    "langgraph>=0.2.64",
    "openai>=1.59.8",
    "streamlit>=1.41.1",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.1",
    "python-dotenv>=1.0.1",
    "ipykernel>=6.29.5",
    "pymongo>=4.11.1",
    "pypdf>=5.3.0",
    "langchain-mongodb>=0.4.0",
]
